+++ code/core/src/main/java/org/betonquest/betonquest/api/quest/QuestTypeApi.java
@@ -17,7 +17,7 @@ import java.util.List;
 public interface QuestTypeApi {
 
     /**
-     * Checks if the conditions described by conditionID are met.
+     * Checks whether all the conditions described by conditionIDs are met.
      *
      * @param profile      the {@link Profile} of the player which should be checked
      * @param conditionIDs IDs of the conditions to check
@@ -26,13 +26,13 @@ public interface QuestTypeApi {
     boolean conditions(@Nullable Profile profile, Collection<ConditionID> conditionIDs);
 
     /**
-     * Checks if the conditions described by conditionID are met.
+     * Checks whether any of the conditions described by conditionIDs are met.
      *
      * @param profile      the {@link Profile} of the player which should be checked
      * @param conditionIDs IDs of the conditions to check
      * @return if all conditions are met
      */
-    boolean conditions(@Nullable Profile profile, ConditionID... conditionIDs);
+    boolean conditionsAny(@Nullable Profile profile, Collection<ConditionID> conditionIDs);
 
     /**
      * Checks if the condition described by conditionID is met.