+++ code/core/src/main/java/org/betonquest/betonquest/api/quest/event/PlayerEvent.java
@@ -2,13 +2,15 @@ package org.betonquest.betonquest.api.quest.event;
 
 import org.betonquest.betonquest.api.QuestException;
 import org.betonquest.betonquest.api.profile.Profile;
+import org.betonquest.betonquest.api.quest.PrimaryThreadEnforceable;
 
 /**
  * Interface for quest-events that are executed for a profile. It represents the player event as described in the
  * BetonQuest user documentation. It does not represent the playerless variant though, see {@link PlayerlessEvent}.
  */
 @FunctionalInterface
-public interface PlayerEvent {
+public interface PlayerEvent extends PrimaryThreadEnforceable {
+
     /**
      * Executes the event.
      *